+++ examples/multilingual/src/main.rs
@@ -13,6 +13,7 @@ fn get_labels_german() -> FileDialogLabels {
     FileDialogLabels {
         title_select_directory: "📁 Ordner Öffnen".to_string(),
         title_select_file: "📂 Datei Öffnen".to_string(),
+        title_select_multiple: "🗐 Mehrere Öffnen".to_string(),
         title_save_file: "📥 Datei Speichern".to_string(),
 
         cancel: "Abbrechen".to_string(),
@@ -39,6 +40,7 @@ fn get_labels_german() -> FileDialogLabels {
 
         selected_directory: "Ausgewählter Ordner:".to_string(),
         selected_file: "Ausgewählte Datei:".to_string(),
+        selected_items: "Ausgewählte Elemente:".to_string(),
         file_name: "Dateiname:".to_string(),
         file_filter_all_files: "Alle Dateien".to_string(),